+++ HISTORY.md
@@ -9,6 +9,7 @@ v1.31.0 ()
 * Support online check ip which used by wwanleds/cloud_watch service
 * Support data cache enable command
 * Fixed file buffer index saving issue
+* Support digest auth method in http.restful
 
 
 v1.30.0 (2021-04-22)